pyproject.toml
[project]
name = "registered"
version = "0.1.0"
description = "A set of scripts for working with TransitMaster"
readme = "README.md"
requires-python = ">=3.10"
packages = ["registered"]
dependencies = [
    "attrs>=24.2.0",
    "folium>=0.18.0",
    "networkx>=2.4",
    "osmnx>=1.2.3",
    "pyinquirer>=1.0.3",
    "pyodbc>=5.2.0",
    "pyproj>=3.6.1; sys_platform != 'win32'",
    "rtree>=1.3.0",
    "shapely<2",
    "smbprotocol>=1.14.0",
]

[tool.uv]
package = false
dev-dependencies = [
    "atomicwrites>=1.4.1",
    "black>=24.10.0",
    "pylint<3",
    "pytest>=8.3.3",
    "pytest-cov>=5.0.0",
    "pytest-watch>=4.2.0",
]

[tool.uv.sources]
networkx = { git = "https://github.com/paulswartz/networkx.git", rev = "sp-2.8" }
